Mini project 3
Uploaded by Hu Yinglun;
The codes for exchange rates somehow did not work and AI also did not fix the problem for me. I manually put in exchange rates in the latte file. 
For the table, pos index means latte in the country is comparatively cheap, while neg index to be expensive.